# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Dutch Market API",
        "version": "1.0.0",
        "description": "Multi-vendor marketplace API. Authenticate with OAuth 2.0 (authorization code + PKCE, client credentials) or a first-party token from `/api/v1/auth/login`."
    },
    "servers": [
        {
            "url": "http://dutchmarket.test/laravel/public/api"
        }
    ],
    "security": [
        {
            "oauth2": []
        }
    ],
    "paths": {
...
        "/v1/orders/{order}/confirm": {
            "post": {
                "operationId": "order.confirm",
                "summary": "Buyer confirms receipt; escrow is released to the vendor",
                "tags": [
                    "Order"
                ],
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "description": "The order ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`OrderResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/OrderResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "items"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/products/{product}": {
            "put": {
                "operationId": "product.update",
                "summary": "Update one of your own products. Requires `catalog:write`",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProductResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "product.destroy",
                "summary": "Delete one of your own products. Requires `catalog:write`",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "http://dutchmarket.test/laravel/public/oauth/authorize",
                        "tokenUrl": "http://dutchmarket.test/laravel/public/oauth/token",
                        "scopes": {
                            "profile": "Read your own account profile",
                            "catalog:read": "Read the product catalog",
                            "catalog:write": "Create, update and delete your own products",
                            "orders:read": "Read your own orders",
                            "orders:write": "Place and manage your own orders",
                            "vendor:manage": "Manage your vendor shop",
                            "admin": "Administer the marketplace"
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "OrderResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "number": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "escrow_status": {
                        "type": "string"
                    },
                    "shipment_status": {
                        "type": "string"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "subtotal_cents": {
                        "type": "integer"
                    },
                    "payment_method": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "tracking_number": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "shipping_address": {
                        "type": "string"
                    },
                    "buyer_id": {
                        "type": "integer"
                    },
                    "vendor_id": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "product_id": {
                                    "type": [
                                        "integer",
                                        "null"
                                    ]
                                },
                                "title": {
                                    "type": "string"
                                },
                                "unit_price_cents": {
                                    "type": "integer"
                                },
                                "quantity": {
                                    "type": "integer"
                                },
                                "line_total_cents": {
                                    "type": "integer"
                                }
                            },
                            "required": [
                                "product_id",
                                "title",
                                "unit_price_cents",
                                "quantity",
                                "line_total_cents"
                            ]
                        }
                    },
                    "paid_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "shipped_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "confirmed_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "number",
                    "status",
                    "escrow_status",
                    "shipment_status",
                    "currency",
                    "subtotal_cents",
                    "payment_method",
                    "tracking_number",
                    "buyer_id",
                    "vendor_id",
                    "paid_at",
                    "shipped_at",
                    "confirmed_at",
                    "created_at"
                ],
                "title": "OrderResource"
            },
            "ProductRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 20000
                    },
                    "price_cents": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 100000000
                    },
                    "currency": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 3
                    },
                    "stock": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 1000000
                    },
                    "category_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "active",
                            "archived"
                        ]
                    }
                },
                "required": [
                    "title",
                    "price_cents"
                ],
                "title": "ProductRequest"
            },
            "ProductResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "price_cents": {
                        "type": "integer"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "vendor": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "handle": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "shop_name": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        },
                        "required": [
                            "id"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "slug",
                    "title",
                    "description",
                    "price_cents",
                    "currency",
                    "stock",
                    "status",
                    "category_id",
                    "vendor",
                    "created_at",
                    "updated_at"
                ],
                "title": "ProductResource"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}